+++ src/routes/venues/proposal/+page.ts
@@ -1,0 +1,13 @@
+import type { PageLoad } from './$types.js';
+
+export const load: PageLoad = async ({ parent, params }) => {
+	const { supabase } = await parent();
+
+	const { data: currencies, error: currenciesError } = await supabase
+		.from('currencies')
+		.select('*');
+
+	return {
+		currencies: currencies && currenciesError === null ? currencies : null
+	};
+};